Split command state and loading
