+++ src/embed-builders/deprecated-prefix.ts
@@ -1,13 +1,10 @@
-import { slashCommandIdsCache } from "@utils/cache";
+import { getSlashCommandMention } from "../state/command-registry";
 import type { Embed } from "lilybird";
 
 export function deprecatedEmbed(commandName: string): Array<Embed.Structure> {
-    const configCommandId = slashCommandIdsCache.get(commandName);
-    const commandMention = configCommandId ?? `/${commandName}`;
-
     return [
         {
-            description: `This prefix command has been deprecated. Use ${commandMention} instead.`,
+            description: `This prefix command has been deprecated. Use ${getSlashCommandMention(commandName)} instead.`,
         },
     ];
 }
